+++ components/openapi/src/test/scala/pl/touk/nussknacker/openapi/http/backend/SharedHttpClientBackendProviderTest.scala
@@ -1,6 +1,10 @@
 package pl.touk.nussknacker.openapi.http.backend
 
+import com.github.tomakehurst.wiremock.WireMockServer
+import com.github.tomakehurst.wiremock.client.WireMock.{get, permanentRedirect, urlEqualTo}
+import com.github.tomakehurst.wiremock.core.WireMockConfiguration
 import org.asynchttpclient.filter.FilterException
+import org.scalatest.BeforeAndAfterAll
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.prop.TableDrivenPropertyChecks
@@ -8,6 +12,7 @@ import pl.touk.nussknacker.engine.api.{JobData, MetaData, ProcessAdditionalField
 import pl.touk.nussknacker.engine.api.runtimecontext.{ContextIdGenerator, EngineRuntimeContext}
 import pl.touk.nussknacker.engine.util.metrics.MetricsProviderForScenario
 import pl.touk.nussknacker.http.backend.DefaultHttpClientConfig
+import pl.touk.nussknacker.test.AvailablePortFinder
 import sttp.client3.basicRequest
 import sttp.model.Uri
 
@@ -15,7 +20,31 @@ import scala.concurrent.Await
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration.DurationInt
 
-class SharedHttpClientBackendProviderTest extends AnyFunSuite with Matchers with TableDrivenPropertyChecks {
+class SharedHttpClientBackendProviderTest
+    extends AnyFunSuite
+    with Matchers
+    with TableDrivenPropertyChecks
+    with BeforeAndAfterAll {
+
+  private val wireMock: WireMockServer = {
+    val server = AvailablePortFinder.withAvailablePortsBlocked(1)(l => {
+      new WireMockServer(
+        WireMockConfiguration
+          .wireMockConfig()
+          .port(l.head)
+      )
+    })
+    server.start()
+    server
+  }
+
+  override protected def afterAll(): Unit = {
+    try {
+      wireMock.stop()
+    } finally {
+      super.afterAll()
+    }
+  }
 
   private val engineRuntimeContext = new EngineRuntimeContext {
     override def jobData: JobData = JobData(
@@ -54,7 +83,6 @@ class SharedHttpClientBackendProviderTest extends AnyFunSuite with Matchers with
     backendProvider.close()
   }
 
-  // todo: lbg
   test("should throw exception on forbidden URL for redirect") {
     val httpConfig = DefaultHttpClientConfig().copy(
       followRedirect = Some(true),
@@ -64,11 +92,17 @@ class SharedHttpClientBackendProviderTest extends AnyFunSuite with Matchers with
     val backendProvider = new SharedHttpClientBackendProvider(httpConfig)
     backendProvider.open(engineRuntimeContext)
 
+    wireMock.stubFor(
+      get(urlEqualTo("/abc")).willReturn(
+        permanentRedirect("http://namespace.nussknacker.svc.cluster.local/abc")
+      )
+    )
+
     val responseFuture = backendProvider.httpBackendForEc
-      .send(basicRequest.get(Uri.unsafeParse("http://namespace.nussknacker.svc.cluster.local/abc")))
+      .send(basicRequest.get(Uri.unsafeParse(s"${wireMock.baseUrl()}/abc")))
 
     intercept[FilterException] {
-      Await.result(responseFuture, 5 seconds)
+      Await.result(responseFuture, 50 seconds)
     }
 
     backendProvider.close()